#20 コードの再整形とディレクトリの移動が完了
